Fix regression
ptgb::vector::resize() had a bug which caused it to incorrectly add/erase elements.

The reason was a loop end condition that was modified during iterations of said loop. This end condition
was intended to be static, but because it used a variable that was modified during the loop, it didn't
repeat the intended number of times.

Fix: before starting the for loop, calculate the end condition once.